Blog -> implements hasPerm

+++ projects/blog/src/blog/Blog.py
@@ -33,22 +33,24 @@ def create(user, title, content):
         db.commit()
         db.close()
 
-def delete(title):
+def delete(user, title):
     if not exists(title):
         print("-- noooo the blog isnt here")
         return False
     else:
+        if hasPerm(user, title):
+            db = sqlite3.connect("/Users/carl/Documents/GitHub/learning-python/projects/blog/blog.db")
+            cursor = db.cursor()
 
-        db = sqlite3.connect("/Users/carl/Documents/GitHub/learning-python/projects/blog/blog.db")
-        cursor = db.cursor()
+            query = "DELETE FROM blog WHERE title = ?"
+            cursor.execute(query, (title,))
+            print("-- blog is DONE! :}")
+            db.commit()
 
-        query = "DELETE FROM blog WHERE title = ?"
-        cursor.execute(query, (title,))
-        print("-- blog is DONE! :}")
-        db.commit()
-
-        db.commit()
-        db.close()
+            db.commit()
+            db.close()
+        else:
+            print("-- no permissions! :}")
 
 def hasPerm(user, title):
     if not exists(title):
